requirements.txt
manimgl==1.7.2
anthropic
google-genai
python-dotenv
pyyaml
edge-tts
